+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/svls/compare/v0.1.17...Unreleased) - ReleaseDate
 
+* [Added] defines config
+
 ## [v0.1.17](https://github.com/dalance/svls/compare/v0.1.16...v0.1.17) - 2020-08-20
 
 * [Changed] Update tower-lsp